Working SLAM
Needs work and cleaning up, filtering is done in a dirty way (deleting a couple of degrees to make slam_toolbox recalculate max). Need to get rid of filtering profile. Need to add localization. WHEN RECORDING A BAG NEED TO CHANGE THE SCAN NUMBER FROM 1600 TO 1599 TO GET RID OF FILTERING HACK.